[DATAB2B-349]Creating a temporary listings relation for agents autonomous (#3168)
* Creating a temporary listings relation for agents autonomous

In this view I'm bringing imovel_aud information to map business migration and partners that type are 'AUTONOMOUS_AGENT'. 
Also excluding an ID used for product tests and previous houses created by partners before the beginning of AA plataform.

* Update bietlejuice/db/dw/queries/datamarts/temp_autonomous_agents_listings.sql

Co-authored-by: Jonas <42319039+vitojon@users.noreply.github.com>

* Update temp_autonomous_agents_listings.sql

* Update temp_autonomous_agents_listings.sql

In order to connect only listings versions after business migration I've changed the rule in the where condition. Now we look for houses created after the agent subscription or, if there is a house migration, the listings with publication after migration.

Co-authored-by: Jonas <42319039+vitojon@users.noreply.github.com>