[DFORSALE 1154]: Add sales_flow_pendency table. (#7604)
Add the data from table sales_flow_pendency on dim_sale_agreement. In order to do that the columns were added to sale_offer and sale_offer_flows. Added on the clean layer queries for sales_flow_pendency and sales_flow_pendency_aud, it also adds it's lineages and updates the lineages for the enrich and dim tables.