[DFORSALE-2424] Remove redundant dependencies in Mediator (#8494)
Redundancies in the dependencies.yaml file happen when one dependency is indirectly satisfied by more than one path.
That is not an error in the first execution of DAGs in the morning. However, if a DAG is triggered during the day, Mediator will trigger DAGs with redundancies more than once, resulting in waste of resources and even errors due to cache.