[DFORSALE-6135] enrich_reports_validation (#12049)
Make reports DAG dependent on reverse_listings_report_access. Why? We want to know the data from the very same day, to compare what was sent to what was received.
Make reports DAG ingest data from the same day, instead of previous day.
Create enrich_reports_validation DAG
Create extra_event table, which identifies any events in the Reports Service that are not in the DW.
Create unprocessed_event table, which identifies any events in the DW that are not in the Reports Service.