[DIN-750] Allow different partitions between raw and clean for Builder (#14797)
Create parameters for raw partitions, clean partitions, raw extraction type and clean extraction type
Create unit tests for TableAttributes
Pass table_customization parameter to raw table in RawDatabasePull. This is to fix an edge case in which the table has cases and is converted to lowercase.
Migrate Kodak DAG to DAG Builder to test this new feature.