🐛 fix gsheets that were having issues (#15796)
* 🐛 fix gsheets that were having issues

* 📝 fix database name in gsheets people metadata

* 🚚  rename employee_country to employee_resident_country in navent_convenia_like.sql

* ♻️  update ts_load column value to current timestamp

* ⚡️ Update navent_convenia_like.sql to include assignment numbers and manager assignment numbers

The code changes in `navent_convenia_like.sql` include modifications to select the `id_colaborador` as `assignment_number` and `id_gestor` as `manager_assignment_number`. Additionally, the `salario_moeda_local` column is split into `currency` and `salary_local_currency` columns. The `status_gestor` column is used to determine if the manager is active. The `dt_inicio`, `dt_desligamento`, and `dt_nascimento` columns are converted to date format.

* ⚡️ Convert ciclo column to uppercase in talent_review.sql

The code changes in `talent_review.sql` include converting the `ciclo` column to uppercase in the SELECT statement. This ensures consistent formatting of the data.

* 📝  Update database names and column names in gsheets people metadata