[DIN-1156] Implement cluster pool (Part 1) (#16119)
Create configs for cluster pools in databricks_10_4_med_general, databricks_12_2_med_general and databricks_13_3_med_general in forno.
Create configs for cluster pools in databricks_10_4_med_general in prod. We intend to release these to the rest tomorrow, this is just a first test with a smaller number of affected clusters.