[DIN-1651] Refactor dag_inventory (#20697)
Change the way we identify which tables are generated by which Airflow tasks. Instead of using sync-metadata, we now use params, which were added in this PR.
Use the CustomIngestionWorkflow, even though it's not actually in the DAG Builder (the .py file is still generated manually). We can't use the DAG Builder and DAG Declarations because dag_inventory runs some custom tasks at the Airflow-level before the Spark job is even triggered on Databricks.
Migrate clean layer to Delta
Upgrade to DBR 16.4